finished Etiquette,Variety and it's impl

+++ src/main/java/com/example/Cataloguemicroservice/Repository/EtiquetteRepository.java
@@ -3,5 +3,5 @@ package com.example.Cataloguemicroservice.Repository;
 import com.example.Cataloguemicroservice.Entities.Etiquette;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EtiquetteRepository extends JpaRepository<Etiquette,Integer> {
+public interface EtiquetteRepository extends JpaRepository<Etiquette,Long> {
 }
